fix circular dependancy created by security groups
We want the internal network to be able to have unfettered network
traffic. In order to provide this, we consolidated all external
traffic down to a single security group and then added an internal
security group. The internal security group allowed any member of
the external security group (i.e., all validators and agents) to
communicate on any port and with any protocol amongst themselves.

Every other method that was tried did not break the dependency cycle
that was created by the last commit.