fix support for using test code
Previous to this, this template, while it had an option to use the
test code, pretty much expected one to be using the stable code.
There were some hard-coded values in some of the post-install
`cloud-init` scripts that did not take the user's selection into
consideration, especially on the agent side.