Update to include correct packages
The 'sovrin-node' and 'sovrin-client' packages have been replaced
in the Sovrin repo with a single 'sovrin' package. Updated scripts
to match. In addition, added missing debsigs, debsig-verify, and
apt-transport-https packages.  In response to user feedback,
modified Vagrantfile to spin up 3 agents and a CLI node by
default, rather than on-demand.